wip: Fix workflow handle logic
