+++ agents-api/agents_api/workflows/embed_docs.py
@@ -9,7 +9,7 @@ with workflow.unsafe.imports_passed_through():
     from ..activities.embed_docs import embed_docs
     from ..activities.types import EmbedDocsPayload
     from ..common.retry_policies import DEFAULT_RETRY_POLICY
-    from ..env import temporal_schedule_to_close_timeout, temporal_heartbeat_timeout
+    from ..env import temporal_heartbeat_timeout, temporal_schedule_to_close_timeout
 
 
 @workflow.defn
